refactor: LottoTicket 을 생성하고, 교환하는 객체들 ticket 패키지 하위로 이동
